De-duplicated code in NavierStokes explicit strategies
